Add local user management and refactor error handling
Implemented passportjs with Mongoose to handle user management in the UI.
Modelled out users to have roles in order to implement access control.
Local and GitHub authentication was implemented as login options

Error handling was switched from being render-based to being
redirect-based. The issue previously was that if displaying an error
on a view that needed some sort of logic/data retrieveal for it, the error
would display on the page but none of the data that needed to be on the page
was displayed. Now all errors are redirected to an endpoint and a flash
message system was implemented leveraging the session store that allows for
displaying error, warning, info, or success messages on any page.